+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 3.4.1
 
+* added tuple for session state cancellation
 * removed the requirement to send a page view before an event can be tracked
 
 ## 3.4.0